+++ Repositories/Interfaces/IShopsRepository.cs
@@ -9,6 +9,6 @@ namespace AmarBariAPI.Repositories.Interfaces
         Task<Result<ShopResponseDto>> GetById(long id);
         Task<Result<long>> Create(ShopRequestDto dto);
         Task<Result<ShopResponseDto>> Update(ShopRequestDto dto);
-        Task<Result<string>> Delete(int id);
+        Task<Result<string>> Delete(long id);
     }
 }
